move formatting function to utils
